Use `arbitraryBuiltin @Integer` to generate quantities in `Value`
